fix Print::printf implementation

+++ cores/nRF5/Print.h
@@ -83,7 +83,7 @@ class Print
     size_t println(double, int = 2);
     size_t println(const Printable&);
     size_t println(void);
-    int    printf(const char* format, ...) __attribute__ ((format (printf, 2, 3)));
+    size_t printf(const char* format, ...) __attribute__ ((format (printf, 2, 3)));
 
     virtual void flush() { /* Empty implementation for backward compatibility */ }
